added check_reader_unreader acceptance test.
