+++ spec/acceptance_helper.rb
@@ -1,0 +1,17 @@
+module AcceptanceHelper
+  def authenticate(name, password)
+    if page.driver.respond_to?(:basic_auth)
+      page.driver.basic_auth(name, password)
+    elsif page.driver.respond_to?(:basic_authorize)
+      page.driver.basic_authorize(name, password)
+    elsif page.driver.respond_to?(:browser) && page.driver.browser.respond_to?(:basic_authorize)
+      page.driver.browser.basic_authorize(name, password)
+    else
+      raise "I don't know how to log in!"
+    end
+  end
+
+  def authenticate_to_admin!
+    authenticate("habitat", "iloveyou")
+  end
+end
