+++ spec/rails_helper.rb
@@ -2,6 +2,7 @@
 ENV["RAILS_ENV"] ||= 'test'
 
 require 'spec_helper'
+require 'acceptance_helper'
 require File.expand_path("../../config/environment", __FILE__)
 require 'rspec/rails'
 
@@ -24,6 +25,8 @@ RSpec.configure do |config|
   config.before(:suite) do
     FactoryGirl.lint
   end
+
+ config.include AcceptanceHelper
 
   # If you're not using ActiveRecord, or you'd prefer not to run each of your
   # examples within a transaction, remove the following line or assign false
